requirements.txt
streamlit>=1.32.0
pdfplumber>=0.11.0
openai>=1.13.0
genanki>=0.13.0
python-dotenv>=1.0.0
# OCR support for scanned PDFs
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0
